[BE] rename: UserContext 패키지 domain으로 이동 #338
